Update card components.js
Changed from ref{} method to state/setState method. An onChange component was added to the Form class to allow React/console to be aware of the string input characters as they're typed. This can be used to suggest stronger passwords. This will likely be useful in the Tanach/Gematria search project.